requirements.txt
faster-whisper
llama-cpp-python
numpy
sounddevice
